Refactored Locks to ConcurrencySupport

+++ tests/src/test/java/org/antublue/verifyica/test/ParallelArgumentSupplierTest1.java
@@ -25,7 +25,6 @@ import org.antublue.verifyica.api.Argument;
 import org.antublue.verifyica.api.ArgumentContext;
 import org.antublue.verifyica.api.ClassContext;
 import org.antublue.verifyica.api.Verifyica;
-import org.antublue.verifyica.engine.support.RandomSupport;
 
 /** Example test */
 public class ParallelArgumentSupplierTest1 {
@@ -79,7 +78,7 @@ public class ParallelArgumentSupplierTest1 {
         assertThat(argumentContext.getStore()).isNotNull();
         assertThat(argumentContext.getTestArgument()).isNotNull();
 
-        Thread.sleep(RandomSupport.randomInt(0, 1000));
+        Thread.sleep(RandomSupport.randomLong(0, 1000));
     }
 
     @Verifyica.Test
@@ -90,7 +89,7 @@ public class ParallelArgumentSupplierTest1 {
         assertThat(argumentContext.getStore()).isNotNull();
         assertThat(argumentContext.getTestArgument()).isNotNull();
 
-        Thread.sleep(RandomSupport.randomInt(0, 1000));
+        Thread.sleep(RandomSupport.randomLong(0, 1000));
     }
 
     @Verifyica.AfterEach
