+++ tests/src/test/java/org/antublue/verifyica/test/argument/AutoCloseableCustomArgumentTest.java
@@ -27,7 +27,7 @@ import org.antublue.verifyica.api.Argument;
 import org.antublue.verifyica.api.ArgumentContext;
 import org.antublue.verifyica.api.ClassContext;
 import org.antublue.verifyica.api.Verifyica;
-import org.antublue.verifyica.engine.support.RandomSupport;
+import org.antublue.verifyica.test.RandomSupport;
 
 /** Example test */
 public class AutoCloseableCustomArgumentTest {
@@ -78,7 +78,7 @@ public class AutoCloseableCustomArgumentTest {
         assertThat(argumentContext.getStore()).isNotNull();
         assertThat(argumentContext.getTestArgument()).isNotNull();
 
-        Thread.sleep(RandomSupport.randomInt(0, 1000));
+        Thread.sleep(RandomSupport.randomLong(0, 1000));
     }
 
     @Verifyica.Test
@@ -89,7 +89,7 @@ public class AutoCloseableCustomArgumentTest {
         assertThat(argumentContext.getStore()).isNotNull();
         assertThat(argumentContext.getTestArgument()).isNotNull();
 
-        Thread.sleep(RandomSupport.randomInt(0, 1000));
+        Thread.sleep(RandomSupport.randomLong(0, 1000));
     }
 
     @Verifyica.AfterEach
