+++ tests/src/test/java/org/antublue/verifyica/test/interceptor/engine/EngineInterceptorTest3.java
@@ -28,7 +28,7 @@ import org.antublue.verifyica.api.Verifyica;
 import org.antublue.verifyica.api.interceptor.engine.ClassDefinition;
 import org.antublue.verifyica.api.interceptor.engine.EngineInterceptor;
 import org.antublue.verifyica.api.interceptor.engine.EngineInterceptorContext;
-import org.antublue.verifyica.engine.support.RandomSupport;
+import org.antublue.verifyica.test.RandomSupport;
 
 /** Example test */
 public class EngineInterceptorTest3 implements EngineInterceptor {
@@ -93,7 +93,7 @@ public class EngineInterceptorTest3 implements EngineInterceptor {
         assertThat(argumentContext.getStore()).isNotNull();
         assertThat(argumentContext.getTestArgument()).isNotNull();
 
-        Thread.sleep(RandomSupport.randomInt(0, 1000));
+        Thread.sleep(RandomSupport.randomLong(0, 1000));
     }
 
     @Verifyica.Test
@@ -104,7 +104,7 @@ public class EngineInterceptorTest3 implements EngineInterceptor {
         assertThat(argumentContext.getStore()).isNotNull();
         assertThat(argumentContext.getTestArgument()).isNotNull();
 
-        Thread.sleep(RandomSupport.randomInt(0, 1000));
+        Thread.sleep(RandomSupport.randomLong(0, 1000));
     }
 
     @Verifyica.AfterEach
